Make waveform render updates smoother, faster and more reliable

+++ src/include/audiotrack.h
@@ -9,7 +9,7 @@ class vsthost;
 struct audiotrack_split_t : public samplesource_t {
     int64_t sampleId = 0;
     int64_t version  = 0;
-    int32_t samplePos;
+    int64_t samplePos;
     audiosample_t sample;
     audiosample_t* getSample() override {
         return &sample;
@@ -25,7 +25,7 @@ struct audiotrack_t {
     std::vector<std::shared_ptr<audiotrack_block_t>> data;
     std::vector<std::shared_ptr<audiotrack_split_t>> samples;
     std::shared_ptr<audiotrack_split_t> getSample(int32_t samplePos);
-    std::shared_ptr<audiotrack_split_t> getSampleById(int32_t sampleId);
+    std::shared_ptr<audiotrack_split_t> getSampleById(int64_t sampleId);
     template<typename Functor>
     void visitSamples(Functor f) {
         ThreadLock lock = MainCtrl::getPlayThread()->lockThread();
